requirements/dev.txt
fastapi==0.88.0
alembic==1.8.1
psycopg2==2.9.3
SQLAlchemy==1.4.45
uvicorn==0.20.0
pyotp==2.7.0
passlib==1.7.4
python-decouple==3.3
python-jose==3.3.0
pydantic==1.10.2
pydantic[email]
ptpython==3.0.20
aiohttp==3.8.3
PyYAML==6.0
aioredis==2.0.1
apscheduler==3.9.1
dateparser==1.1.8